Calculate the Forest OzCBI from numeric input variables

calculate_ozcbi_forest(
  stratum_1_surface_p_s1_unburnt = NA_real_,
  stratum_1_surface_p_s1_duff = NA_real_,
  stratum_3_elevated_p_s3_crown = NA_real_,
  straum_4_intermediate_p_s4_crown = NA_real_,
  straum_4_intermediate_p_s4_char = NA_real_,
  stratum_5_overstorey_p_s5_crown = NA_real_,
  stratum_5_overstorey_p_s5_litter = NA_real_,
  stratum_5_overstorey_p_s5_char = NA_real_,
  stratum_1_surface_s1_fcov = 1,
  stratum_3_elevated_p_s3_fcov = 0,
  straum_4_intermediate_p_s4_fcov = 0,
  stratum_5_overstorey_p_s5_fcov = 0,
  verbose = FALSE
)

Arguments

stratum_1_surface_p_s1_unburnt

A numeric grade from 0.0 to 3.0. Unburnt area. The approximate percentage of area not burned.

stratum_1_surface_p_s1_duff

A numeric grade from 0.0 to 3.0. Duff condition (crushed sticks and leaves). Broken leaf pieces that form a type of mulch under the litter layer. If necessary, scrape down to mineral soil to see how deeply the char has penetrated.

stratum_3_elevated_p_s3_crown

A numeric grade from 0.0 to 3.0. Intact original crown cover. How much original crown is intact?

straum_4_intermediate_p_s4_crown

A numeric grade from 0.0 to 3.0. Intact original crown cover. How much original crown is intact?

straum_4_intermediate_p_s4_char

A numeric grade from 0.0 to 3.0. Char height. Fraction of total stratum height charred.

stratum_5_overstorey_p_s5_crown

A numeric grade from 0.0 to 3.0. Intact original crown cover. How much original crown is intact?

stratum_5_overstorey_p_s5_litter

A numeric grade from 0.0 to 3.0. Ground surface covered by leaves that have fallen after the burn, not unburned patches.

stratum_5_overstorey_p_s5_char

A numeric grade from 0.0 to 3.0. Char height. Fraction of total stratum height charred.

stratum_1_surface_s1_fcov

The fraction of coverage of stratum 1, default: 1. This value is never captured in the digital form as it always is 1 (100%). The variable is however provided here to allow different values.

stratum_3_elevated_p_s3_fcov

The fraction of coverage of stratum 2 in quarter steps from 0.0 to 1.0. Default: 0.

straum_4_intermediate_p_s4_fcov

The fraction of coverage of stratum 2 in quarter steps from 0.0 to 1.0. Default: 0.

stratum_5_overstorey_p_s5_fcov

The fraction of coverage of stratum 2 in quarter steps from 0.0 to 1.0. Default: 0.

verbose

Whether to display diagnostic messages, default: FALSE.

Details

The Forest OzCBI is calculated from the following components:

Stratum 1 fraction of cover: always 1 (100%), therefore never captured.

Stratum 2..5 fcov: surface covered by the stratum in quarter steps from 0 to 1: 0.0, 0.25, 0.5, 0.75, 1.0

Stratum scores (CBI): The average grading from all non-null gradings out of all captured variables for each stratum. The variables are named with a stratum prefix. E.g., all variables for stratum 1 are prefixed stratum_1_surface_s1_.

Overall index: OzCBI

The OzCBI is calculated as:

(sum of stratum scores) / (sum of stratum fraction of cover)

Variables not used in OzCBI formula

Some variables are captured, but not used for OzCBI calculation. They provide context and metadata, and include representative photos.

See also

Examples

# With missing variables
calculate_ozcbi_forest(
  stratum_1_surface_p_s1_unburnt = 1,
  verbose = TRUE
)
#> [1] 1

# With complete variables, all set to 1
calculate_ozcbi_forest(
  stratum_1_surface_p_s1_unburnt = 1,
  stratum_1_surface_p_s1_duff = 1,
  stratum_3_elevated_p_s3_crown = 1,
  straum_4_intermediate_p_s4_crown = 1,
  straum_4_intermediate_p_s4_char = 1,
  stratum_5_overstorey_p_s5_crown = 1,
  stratum_5_overstorey_p_s5_litter = 1,
  stratum_5_overstorey_p_s5_char = 1,
  stratum_1_surface_s1_fcov = 1,
  stratum_3_elevated_p_s3_fcov = 0,
  straum_4_intermediate_p_s4_fcov = 0,
  stratum_5_overstorey_p_s5_fcov = 0,
  verbose = TRUE
)
#> [1] 1